requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py@e62473a1b40177424221bbce02e56db37792c039
async-timeout==3.0.1
beautifulsoup4==4.9.3
lxml==4.6.3
aiohttp==3.7.4.post0
pymongo[srv]
motor==2.5.1
python-dateutil==2.8.2
